requirements.txt
certifi==2020.4.5.1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-heroku==0.1.9
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.1
greenlet==1.0.0
gunicorn==20.0.4
-e git+https://github.com/Nick-palmar/inventory-app.git@2985df7ec1a83bdcc6740d134771f350ef1bb09a#egg=inventory_app
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.10.0
psycopg2==2.8.6
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.4.0
Werkzeug==1.0.1
